refactor: change GDino and GSam services to Agents
